requirements.txt
streamlit
pyfhel
phe
pycryptodome
